changed to parseFloat

+++ main.js
@@ -6,24 +6,21 @@ document.getElementById('divide').addEventListener('click', divide)
 
 function add(){
     let num1 = document.getElementById('number1').value
-    let num2 = document.getElementById('number2').value
-    
+    let num2 = document.getElementById('number2').value    
     document.getElementById('calculation').innerText = parseFloat(Number(num1) + Number(num2))
 }
 function sub(){
     let num1 = document.getElementById('number1').value
     let num2 = document.getElementById('number2').value
-    document.getElementById('calculation').innerText = parseInt(num1 - num2)
+    document.getElementById('calculation').innerText = parseFloat(num1 - num2)
 }
-
 function mult(){
     let num1 = document.getElementById('number1').value
     let num2 = document.getElementById('number2').value
-    document.getElementById('calculation').innerText = parseInt(num1 * num2)
+    document.getElementById('calculation').innerText = parseFloat(num1 * num2)
 }
-
 function divide(){
     let num1 = document.getElementById('number1').value
     let num2 = document.getElementById('number2').value
-    document.getElementById('calculation').innerText = parseInt(num1 / num2)
+    document.getElementById('calculation').innerText = parseFloat(num1 / num2)
 }
